almost there

+++ src/edu/grinnell/csc207/huann/hw2/TwoB.java
@@ -4,8 +4,9 @@ public class TwoB {
 	/**
 	 * Compute the average of two integers.  Rounds toward zero if the
 	 * average is not a whole number.
+	 * Method revised to account for very large int sums.
 	 */
 	public static int average(int left, int right) {
-		return (left + right) / 2;
+		return (int) ((double) (left) + (double) (right)) / 2;
 	} // average(int,int)
 } // class TwoB
